Moved non setup code from core/__init__.py to the tornado.py module.
Refs: #109.
