+++ firenado/components/firenado/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Copyright 2015 Flavio Garcia
+# Copyright 2015-2016 Flavio Garcia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,16 +13,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4:
 
-import firenado.core
+import firenado.tornado
 from firenado.components.firenado.handlers import AppInfoHandler
 
 
-class FirenadoComponent(firenado.core.TornadoComponent):
+class FirenadoComponent(firenado.tornado.TornadoComponent):
 
     def get_handlers(self):
         return [
-            (r'/app/info', AppInfoHandler),
+            (r'/firenado/app/info', AppInfoHandler),
         ]
