+++ firenado/components/firenado/handlers.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Copyright 2015 Flavio Garcia
+# Copyright 2015-2016 Flavio Garcia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,16 +13,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4:
 
 import firenado
 import firenado.conf
-import firenado.core
+import firenado.tornado
 import tornado
 
 
-class AppInfoHandler(firenado.core.TornadoHandler):
+class AppInfoHandler(firenado.tornado.TornadoHandler):
 
     def get(self):
         self.render(
