+++ firenado/core/websocket.py
@@ -19,7 +19,7 @@
 from __future__ import (absolute_import, division,
                         print_function, with_statement)
 
-from firenado.core import JSONError
+from firenado.tornado import JSONError
 from firenado.core import session
 from firenado.core import template
 import tornado.websocket
