+++ firenado/examples/chatdemo/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Copyright 2015 Flavio Garcia
+# Copyright 2015-2016 Flavio Garcia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,11 +13,9 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4:
 
 import chatdemo.handlers
-import firenado.core
+import firenado.tornado
 import logging
 from tornado.concurrent import Future
 
@@ -62,7 +60,7 @@ class MessageBuffer(object):
             self.cache = self.cache[-self.cache_size:]
 
 
-class ChatdemoComponent(firenado.core.TornadoComponent):
+class ChatdemoComponent(firenado.tornado.TornadoComponent):
 
     def __init__(self, name, application, config={}):
         self.message_buffer = MessageBuffer()
