+++ firenado/examples/chatdemo/handlers.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Copyright 2015 Flavio Garcia
+# Copyright 2015-2016 Flavio Garcia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,16 +13,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4:
 
-import firenado.core
+import firenado.tornado
 import tornado.escape
 from tornado import gen
 import uuid
 
 
-class MainHandler(firenado.core.TornadoHandler):
+class MainHandler(firenado.tornado.TornadoHandler):
 
     def get(self):
         print self.component.message_buffer
@@ -30,7 +28,7 @@ class MainHandler(firenado.core.TornadoHandler):
                     messages=self.component.message_buffer.cache)
 
 
-class MessageNewHandler(firenado.core.TornadoHandler):
+class MessageNewHandler(firenado.tornado.TornadoHandler):
 
     def post(self):
         message = {
@@ -48,7 +46,7 @@ class MessageNewHandler(firenado.core.TornadoHandler):
         self.component.message_buffer.new_messages([message])
 
 
-class MessageUpdatesHandler(firenado.core.TornadoHandler):
+class MessageUpdatesHandler(firenado.tornado.TornadoHandler):
 
     @gen.coroutine
     def post(self):
