+++ firenado/examples/skell/__init__.py
@@ -13,15 +13,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4:
 
 import skell.handlers
-import firenado.core
+import firenado.tornado
 from skell import uimodules
 
 
-class SkellComponent(firenado.core.TornadoComponent):
+class SkellComponent(firenado.tornado.TornadoComponent):
 
     def get_handlers(self):
         return [
