+++ firenado/examples/skell/handlers.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 #
-# Copyright 2015 Flavio Garcia
+# Copyright 2015-2016 Flavio Garcia
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,19 +13,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-# vim: set expandtab tabstop=4 shiftwidth=4 softtabstop=4:
 
-import firenado.core
+import firenado.tornado
 
 
-class IndexHandler(firenado.core.TornadoHandler):
+class IndexHandler(firenado.tornado.TornadoHandler):
 
     def get(self):
         self.render("index.html", message="Hello world!!!")
 
 
-class SessionHandler(firenado.core.TornadoHandler):
+class SessionHandler(firenado.tornado.TornadoHandler):
 
     def get(self):
         reset = self.get_argument("reset", False, True)
